+++ Practica 2/plantillaSDL/HolaSDL/Enemy.h
@@ -27,7 +27,7 @@ private:
 	Game* game = nullptr;		// puntero al juego
 
 	// MOVIMIENTO
-	Point2D<int> position;	// posicion actual en Point2D
+	Point2D<double> position;	// posicion actual en Point2D
 	Vector2D<int> direction;	// direccion de movimiento
 
 	int enemyFrame;
@@ -43,7 +43,7 @@ private:
 	SDL_Rect destRect;
 
 public:
-	Enemy(Game* g, Point2D<int> pos, Texture* t);
+	Enemy(Game* g, Point2D<double> pos, Texture* t);
 
 	// -- render --
 	virtual void render() const override;
