+++ Practica 2/plantillaSDL/HolaSDL/Koopa.h
@@ -26,7 +26,7 @@ private:
 
 
 public:
-	Koopa(Game* g, Point2D<int> pos, Texture* texture);
+	Koopa(Game* g, Point2D<double> pos, Texture* texture);
 
 	// -- update --
 	void update();
