+++ Practica 2/plantillaSDL/HolaSDL/Mushroom.h
@@ -24,7 +24,7 @@ private:
 	Game* game = nullptr;		// puntero al juego
 
 	// MOVIMIENTO
-	Point2D<int> position;	// posicion actual en Point2D
+	Point2D<double> position;	// posicion actual en Point2D
 	Vector2D<int> direction;	// direccion de movimiento
 
 	Point2D<int> new_position;	// posicion que tendria si avanza en Point2D
@@ -35,7 +35,7 @@ private:
 	bool alive;
 
 public:
-	Mushroom(Game* g, Point2D<int> pos, Texture* texture);
+	Mushroom(Game* g, Point2D<double> pos, Texture* texture);
 
 	// -- update --
 	void update() override;
